Fix clippy warnings across frontend crates
Burn the five frontend crates' clippy output to a clean `-D warnings` gate
(default features, plus the wpe-webkit variant for renderer-wgpu and
display-runtime). Mechanical lints are fixed in place; bulk categories
(too_many_arguments) and judgment calls (large_enum_variant, type_complexity,
should_implement_trait, result_large_err, enum_variant_names, the FFI proc /
literal-unwrap idioms) get crate-, module-, or item-scoped allows with
one-line justifications. In the wpe-gated backend, genuinely dead items (the
unwired WebKit error-reporting helpers and the unused WPEBuffer import
converters) are deleted, while FFI handle mirrors and RAII lifetime fields are
kept behind justified `#[allow(dead_code)]`. Missing `# Safety` docs and a
stale doc-comment/orphaned import are cleaned up. Behavior-preserving.

+++ neomacs-layout-engine/src/display_row_walk_state.rs
@@ -432,7 +432,7 @@ impl WordWrapRenderState {
 impl HorizontalScrollSkipState {
     pub(crate) fn new(wrap_mode: LineWrapMode, hscroll_columns: i32) -> Self {
         let configured_columns = if wrap_mode == LineWrapMode::Truncate {
-            hscroll_columns.max(0) as i32
+            hscroll_columns.max(0)
         } else {
             0
         };
